requirements.txt
cryptography
flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
httpie
passlib
pymysql
python-dotenv